BlastUnitWorkingData is now its own class
