+++ Source/Libraries/CorruptCore/BlastUnitWorkingData.cs
@@ -1,0 +1,29 @@
+namespace RTCV.CorruptCore
+{
+    using System;
+    using System.Collections.Generic;
+    using Ceras;
+
+    /// <summary>
+    /// Working data for BlastUnits.
+    /// Not serialized
+    /// </summary>
+    [Ceras.MemberConfig(TargetMember.None)]
+    public class BlastUnitWorkingData
+    {
+        //We Calculate a LastFrame at the beginning of execute
+        [NonSerialized]
+        public int LastFrame = -1;
+        //We calculate ExecuteFrameQueued which is the ExecuteFrame + the currentframe that was calculated at the time of it entering the execution pool
+        [NonSerialized]
+        public int ExecuteFrameQueued = 0;
+
+        //We use ApplyValue so we don't need to keep re-calculating the tiled value every execute if we don't have to.
+        [NonSerialized]
+        public byte[] ApplyValue = null;
+
+        //The data that has been backed up. This is a list of bytes so if they start backing up at IMMEDIATE, they can have historical backups
+        [NonSerialized]
+        public Queue<byte[]> StoreData = new Queue<byte[]>();
+    }
+}
